% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxMin_rcpp}
\alias{maxMin_rcpp}
\title{maxMin_rcpp
Funcion para obtener el valor maximo de los valores minimos
entre las columnas de una matriz y las filas de otra matriz, donde
ambas son matrices cuadraticas}
\arguments{
\item{NumericMatriz}{de tipo nxn.}
}
\description{
maxMin_rcpp
Funcion para obtener el valor maximo de los valores minimos
entre las columnas de una matriz y las filas de otra matriz, donde
ambas son matrices cuadraticas
}
\examples{
set.seed(123)
matrixA <- matrix(round(runif(3^2, 0, 1), 1), nrow = 3, ncol = 3)
matrixB <- matrix(round(runif(3^2, 0, 1), 1), nrow = 3, ncol = 3) 
 
matrixA
matrixB

values <- maxMin_rcpp(matrixA, matrixB)
values
}
