% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multMatrix_rcpp}
\alias{multMatrix_rcpp}
\title{multMatrix_rcpp
Funcion para multiplicar dos matrices}
\arguments{
\item{NumericMatriz}{de tipo mxn.}
}
\description{
multMatrix_rcpp
Funcion para multiplicar dos matrices
}
\examples{
set.seed(123)
matrixA <- matrix(round(runif(3^2, 0, 1), 1), nrow = 3, ncol = 3)
matrixB <- matrix(round(runif(3^2, 0, 1), 1), nrow = 3, ncol = 3) 
 
matrixA
matrixB

newMatrix <- multMatrix_rcpp(matrixA, matrixB)
newMatrix
}
