% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichMin_rcpp}
\alias{whichMin_rcpp}
\title{whichMin_rcpp
Funcion para encontrar el/los valores minimos de una matriz,
un data frame con las posiciones}
\arguments{
\item{NumericMatriz}{de tipo nxm.}
}
\description{
whichMin_rcpp
Funcion para encontrar el/los valores minimos de una matriz,
un data frame con las posiciones
}
\examples{
set.seed(123)
matrixA <- matrix(round(runif(3^2, 0, 1), 1), nrow = 3, ncol = 3)
 
matrixA

minValuesPositions <- whichMin_rcpp(matrixA)
minValuesPositions
}
